requirements.txt
cycler==0.10.0
Django==1.11
django-unixtimestampfield==0.3.9
djangorestframework==3.8.2
kiwisolver==1.0.1
matplotlib==3.0.0
numpy==1.15.2
pandas==0.23.4
pyparsing==2.2.2
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
